renames dbt_runner to dbt
